+++ app/routes/chats+/$season.tsx
@@ -23,15 +23,17 @@ import {getSeasonListItems} from '~/utils/simplecast.server.ts'
 import {getServerTimeHeader} from '~/utils/timing.server.ts'
 
 export const handle: KCDHandle = {
-  getSitemapEntries: async request => {
-    const seasons = await getSeasonListItems({request})
-    return seasons.map(season => {
-      return {
-        route: `/chats/${season.seasonNumber.toString().padStart(2, '0')}`,
-        priority: 0.4,
+  getSitemapEntries: import.meta.env.SSR
+    ? async request => {
+        const seasons = await getSeasonListItems({request})
+        return seasons.map(season => {
+          return {
+            route: `/chats/${season.seasonNumber.toString().padStart(2, '0')}`,
+            priority: 0.4,
+          }
+        })
       }
-    })
-  },
+    : null,
 }
 
 type LoaderData = {
