+++ app/utils/search.server.ts
@@ -2,7 +2,7 @@ import {matchSorter, rankings} from 'match-sorter'
 import {getEpisodePath as getCKEpisodePath} from '~/utils/call-kent.ts'
 import {getCWKEpisodePath} from '~/utils/chats-with-kent.ts'
 import {stripHtml} from '~/utils/markdown.server.ts'
-import {getBlogMdxListItems} from '~/utils/mdx.tsx'
+import {getBlogMdxListItems} from '~/utils/mdx.server.tsx'
 import {typedBoolean} from '~/utils/misc.tsx'
 import {getSeasons as getChatsWithKentSeasons} from '~/utils/simplecast.server.ts'
 import {getTalksAndTags} from '~/utils/talks.server.ts'
