+++ package.json
@@ -27,7 +27,7 @@
     "clean": "rimraf ./node_modules/.cache ./server/dist ./build ./public/build",
     "prebuild": "npm run clean && echo All clean ✨",
     "build": "run-s build:*",
-    "build:remix": "remix build --sourcemap",
+    "build:remix": "vite build && vite build --ssr",
     "build:server": "tsx ./other/build-server.ts",
     "build:info": "tsx ./other/generate-build-info",
     "dev": "remix dev -c \"node ./server/dev-server.js\" --manual",
@@ -70,7 +70,6 @@
     "@reach/tabs": "^0.18.0",
     "@remark-embedder/core": "^3.0.2",
     "@remark-embedder/transformer-oembed": "^3.0.0",
-    "@remix-run/css-bundle": "2.3.0",
     "@remix-run/express": "2.3.0",
     "@remix-run/node": "2.3.0",
     "@remix-run/react": "2.3.0",
@@ -132,7 +131,6 @@
     "prisma": "^5.3.1",
     "qrcode": "^1.5.3",
     "react": "^18.2.0",
-    "react-countdown-hook": "^1.1.3",
     "react-dom": "^18.2.0",
     "react-error-boundary": "^4.0.11",
     "react-is": "^18.2.0",
@@ -212,7 +210,9 @@
     "rimraf": "^5.0.1",
     "start-server-and-test": "^2.0.0",
     "tsx": "^4.6.2",
-    "typescript": "^5.2.2"
+    "typescript": "^5.2.2",
+    "vite": "^4.5.0",
+    "vite-tsconfig-paths": "^4.2.1"
   },
   "eslintIgnore": [
     "node_modules",
