+++ tsconfig.json
@@ -6,9 +6,9 @@
     "isolatedModules": true,
     "esModuleInterop": true,
     "jsx": "react-jsx",
-    "module": "nodenext",
+    "module": "ESNext",
     "target": "ES2022",
-    "moduleResolution": "nodenext",
+    "moduleResolution": "Bundler",
     "resolveJsonModule": true,
     "noUncheckedIndexedAccess": true,
     "strict": true,
